requirements.txt
# requirements.txt
Flask
boto3
google-cloud-storage
python-dotenv